# Do not generate error messages in TRY_CAST and TRY(CAST)