--- a/velox/expression/TryExpr.cpp
+++ b/velox/expression/TryExpr.cpp
@@ -26,6 +26,8 @@ void TryExpr::evalSpecialForm(
   ScopedVarSetter captureErrorDetails(
       context.mutableCaptureErrorDetails(), false);
 
+  ScopedThreadSkipErrorDetails skipErrorDetails(true);
+
   // It's possible with nested TRY expressions that some rows already threw
   // exceptions in earlier expressions that haven't been handled yet. To avoid
   // incorrectly handling them here, store those errors and temporarily reset
@@ -47,6 +49,8 @@ void TryExpr::evalSpecialFormSimplified(
   ScopedVarSetter throwOnError(context.mutableThrowOnError(), false);
   ScopedVarSetter captureErrorDetails(
       context.mutableCaptureErrorDetails(), false);
+
+  ScopedThreadSkipErrorDetails skipErrorDetails(true);
 
   // It's possible with nested TRY expressions that some rows already threw
   // exceptions in earlier expressions that haven't been handled yet. To avoid