--- a/velox/type/Conversions.h
+++ b/velox/type/Conversions.h
@@ -70,6 +70,9 @@ template <typename T, typename F>
 Expected<T> callFollyTo(const F& v) {
   const auto result = folly::tryTo<T>(v);
   if (result.hasError()) {
+    if (threadSkipErrorDetails()) {
+      return folly::makeUnexpected(Status::UserError(""));
+    }
     return folly::makeUnexpected(Status::UserError(
         "{}", folly::makeConversionError(result.error(), "").what()));
   }